test: refactor TransactionViewModelTest into domain-specific classes
